rove: restructure crate
